Move alliance in separate folder
