+++ src/components/alliance/allianceItemData.ts
